Extract record storage test base, make stand storage tests use separate test entity class
