+++ server/src/test/java/io/spine/server/storage/RecordStorageTest.java
@@ -20,11 +20,8 @@
 
 package io.spine.server.storage;
 
-import com.google.common.base.Function;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Maps;
 import com.google.protobuf.Any;
-import com.google.protobuf.Descriptors;
 import com.google.protobuf.FieldMask;
 import com.google.protobuf.Int32Value;
 import com.google.protobuf.Message;
@@ -38,8 +35,9 @@ import io.spine.core.Version;
 import io.spine.protobuf.TypeConverter;
 import io.spine.server.entity.Entity;
 import io.spine.server.entity.EntityRecord;
-import io.spine.server.entity.FieldMasks;
 import io.spine.server.entity.LifecycleFlags;
+import io.spine.server.entity.TestTransaction;
+import io.spine.server.entity.TransactionalEntity;
 import io.spine.server.entity.storage.EntityColumn;
 import io.spine.server.entity.storage.EntityColumn.MemoizedValue;
 import io.spine.server.entity.storage.EntityQuery;
@@ -47,16 +45,12 @@ import io.spine.server.entity.storage.EntityRecordWithColumns;
 import io.spine.server.storage.given.RecordStorageTestEnv.LifecycleColumns;
 import io.spine.server.storage.given.RecordStorageTestEnv.TestCounterEntity;
 import io.spine.test.storage.Project;
-import io.spine.testdata.Sample;
+import io.spine.test.storage.ProjectId;
 import io.spine.testing.core.given.GivenVersion;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import java.util.Collection;
-import java.util.HashMap;
 import java.util.Iterator;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -64,7 +58,7 @@ import java.util.Set;
 
 import static com.google.common.collect.Lists.newArrayList;
 import static com.google.common.collect.Sets.newHashSet;
-import static io.spine.base.Identifier.pack;
+import static io.spine.base.Identifier.newUuid;
 import static io.spine.client.ColumnFilters.all;
 import static io.spine.client.ColumnFilters.eq;
 import static io.spine.client.CompositeColumnFilter.CompositeOperator.ALL;
@@ -76,71 +70,33 @@ import static io.spine.server.entity.storage.TestEntityRecordWithColumnsFactory.
 import static io.spine.server.storage.LifecycleFlagField.archived;
 import static io.spine.test.storage.Project.Status.CANCELLED;
 import static io.spine.test.storage.Project.Status.DONE;
-import static io.spine.testing.Tests.assertMatchesMask;
-import static io.spine.testing.Verify.assertIteratorsEqual;
-import static io.spine.testing.Verify.assertSize;
-import static io.spine.testing.server.entity.given.GivenLifecycleFlags.archived;
-import static io.spine.validate.Validate.isDefault;
-import static java.util.stream.Collectors.toList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
-import static org.mockito.ArgumentMatchers.argThat;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 /**
  * @author Dmytro Dashenkov
  */
 @SuppressWarnings("unused") // JUnit nested classes considered unused in abstract class.
-public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
-        extends AbstractStorageTest<I, EntityRecord, RecordReadRequest<I>, S> {
-
-    private static EntityRecord newStorageRecord(Message state) {
-        Any wrappedState = pack(state);
-        EntityRecord record = EntityRecord.newBuilder()
-                                          .setState(wrappedState)
-                                          .setVersion(GivenVersion.withNumber(0))
-                                          .build();
-        return record;
-    }
-
-    /**
-     * Creates an unique {@code Message} with the specified ID.
-     *
-     * <p>Two calls for the same ID should return messages, which are not equal.
-     *
-     * @param id the ID for the message
-     * @return the unique {@code Message}
-     */
-    protected abstract Message newState(I id);
-
-    @Override
-    protected EntityRecord newStorageRecord() {
-        return newStorageRecord(newState(newId()));
-    }
-
-    @Override
-    protected RecordReadRequest<I> newReadRequest(I id) {
-        return new RecordReadRequest<>(id);
-    }
+public abstract class RecordStorageTest<S extends RecordStorage<ProjectId>>
+        extends AbstractRecordStorateTest<ProjectId, S> {
 
     @Override
     protected Class<? extends TestCounterEntity> getTestEntityClass() {
         return TestCounterEntity.class;
     }
 
-    private EntityRecord newStorageRecord(I id) {
-        return newStorageRecord(newState(id));
+    @Override
+    protected final ProjectId newId() {
+        return ProjectId.newBuilder()
+                        .setId(newUuid())
+                        .build();
     }
 
-    private EntityRecord newStorageRecord(I id, Message state) {
+    private static EntityRecord newStorageRecord(ProjectId id, Message state) {
         Any wrappedState = pack(state);
         EntityRecord record = EntityRecord
                 .newBuilder()
@@ -151,288 +107,32 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
         return record;
     }
 
-    @SuppressWarnings("ConstantConditions")
-    // Converter nullability issues and Optional getting
-    @Test
-    @DisplayName("write and read record by Message ID")
-    void writeAndReadByMessageId() {
-        RecordStorage<I> storage = getStorage();
-        I id = newId();
-        EntityRecord expected = newStorageRecord(id);
-        storage.write(id, expected);
-
-        RecordReadRequest<I> readRequest = newReadRequest(id);
-        EntityRecord actual = storage.read(readRequest)
-                                     .get();
-
-        assertEquals(expected, actual);
-        close(storage);
+    /**
+     * Creates new instance of the test entity.
+     */
+    private static TestCounterEntity newEntity(ProjectId id) {
+        return new TestCounterEntity(id);
     }
 
-    @Test
-    @DisplayName("retrieve empty iterator if storage is empty")
-    void retrieveEmptyIterator() {
-        FieldMask nonEmptyFieldMask = FieldMask.newBuilder()
-                                               .addPaths("invalid-path")
-                                               .build();
-        RecordStorage storage = getStorage();
-        Iterator empty = storage.readAll(nonEmptyFieldMask);
-
-        assertNotNull(empty);
-        assertFalse(empty.hasNext(), "Iterator is not empty!");
+    private static void archive(TransactionalEntity<ProjectId, ?, ?> entity) {
+        TestTransaction.archive(entity);
     }
 
-    @Nested
-    @DisplayName("given field mask, read")
-    class ReadWithMask {
-
-        @SuppressWarnings("ConstantConditions") // Converter nullability issues.
-        @Test
-        @DisplayName("single record")
-        void singleRecord() {
-            I id = newId();
-            EntityRecord record = newStorageRecord(id);
-            RecordStorage<I> storage = getStorage();
-            storage.write(id, record);
-
-            Descriptors.Descriptor descriptor = newState(id).getDescriptorForType();
-            FieldMask idMask = FieldMasks.maskOf(descriptor, 1);
-
-            RecordReadRequest<I> readRequest = new RecordReadRequest<>(id);
-            Optional<EntityRecord> optional = storage.read(readRequest, idMask);
-            assertTrue(optional.isPresent());
-            EntityRecord entityRecord = optional.get();
-
-            Message unpacked = unpack(entityRecord.getState());
-            assertFalse(isDefault(unpacked));
-        }
-
-        @SuppressWarnings({"MethodWithMultipleLoops", "ConstantConditions"})
-        // Converter nullability issues.
-        @Test
-        @DisplayName("multiple records")
-        void multipleRecords() {
-            RecordStorage<I> storage = getStorage();
-            int count = 10;
-            List<I> ids = new LinkedList<>();
-            Descriptors.Descriptor typeDescriptor = null;
-
-            for (int i = 0; i < count; i++) {
-                I id = newId();
-                Message state = newState(id);
-                EntityRecord record = newStorageRecord(state);
-                storage.write(id, record);
-                ids.add(id);
-
-                if (typeDescriptor == null) {
-                    typeDescriptor = state.getDescriptorForType();
-                }
-            }
-
-            int bulkCount = count / 2;
-            FieldMask fieldMask = FieldMasks.maskOf(typeDescriptor, 2);
-            Iterator<EntityRecord> readRecords = storage.readMultiple(
-                    ids.subList(0, bulkCount),
-                    fieldMask);
-            List<EntityRecord> readList = newArrayList(readRecords);
-            assertSize(bulkCount, readList);
-            for (EntityRecord record : readList) {
-                Message state = unpack(record.getState());
-                assertMatchesMask(state, fieldMask);
-            }
-        }
-    }
-
-    @SuppressWarnings("ConstantConditions") // converter nullability issues
-    @Test
-    @DisplayName("delete record")
-    void deleteRecord() {
-        RecordStorage<I> storage = getStorage();
-        I id = newId();
-        EntityRecord record = newStorageRecord(id);
-
-        // Write the record.
-        storage.write(id, record);
-
-        // Delete the record.
-        assertTrue(storage.delete(id));
-
-        // There's no record with such ID.
-        RecordReadRequest<I> readRequest = newReadRequest(id);
-        assertFalse(storage.read(readRequest)
-                           .isPresent());
-    }
-
-    @Test
-    @DisplayName("write none storage fields if none are passed")
-    void writeNoneFieldsIfNonePassed() {
-        RecordStorage<I> storage = spy(getStorage());
-        I id = newId();
-        Any state = pack(Sample.messageOfType(Project.class));
-        EntityRecord record =
-                Sample.<EntityRecord, EntityRecord.Builder>builderForType(EntityRecord.class)
-                        .setState(state)
-                        .build();
-        storage.write(id, record);
-        verify(storage).write(eq(id), withRecordAndNoFields(record));
-    }
-
-    @Nested
-    @DisplayName("given bulk of records, write them")
-    class WriteBulk {
-
-        @Test
-        @DisplayName("for the first time")
-        void forTheFirstTime() {
-            RecordStorage<I> storage = getStorage();
-            int bulkSize = 5;
-
-            Map<I, EntityRecordWithColumns> initial = new HashMap<>(bulkSize);
-
-            for (int i = 0; i < bulkSize; i++) {
-                I id = newId();
-                EntityRecord record = newStorageRecord(id);
-                initial.put(id, EntityRecordWithColumns.of(record));
-            }
-            storage.write(initial);
-
-            Collection<EntityRecord> actual = newArrayList(
-                    storage.readMultiple(initial.keySet())
-            );
-
-            Collection<EntityRecord> expected =
-                    initial.values()
-                           .stream()
-                           .map(recordWithColumns -> recordWithColumns != null
-                                                     ? recordWithColumns.getRecord()
-                                                     : null)
-                           .collect(toList());
-
-            assertEquals(expected.size(), actual.size());
-            assertTrue(actual.containsAll(expected));
-
-            close(storage);
-        }
-
-        @Test
-        @DisplayName("re-writing existing ones")
-        void rewritingExisting() {
-            int recordCount = 3;
-            RecordStorage<I> storage = getStorage();
-
-            Function<EntityRecord, EntityRecordWithColumns> recordPacker =
-                    record -> record != null
-                              ? withLifecycleColumns(record)
-                              : null;
-            Map<I, EntityRecord> v1Records = new HashMap<>(recordCount);
-            Map<I, EntityRecord> v2Records = new HashMap<>(recordCount);
-
-            for (int i = 0; i < recordCount; i++) {
-                I id = newId();
-                EntityRecord record = newStorageRecord(id);
-
-                // Some records are changed and some are not
-                EntityRecord alternateRecord = (i % 2 == 0)
-                                               ? record
-                                               : newStorageRecord(id);
-                v1Records.put(id, record);
-                v2Records.put(id, alternateRecord);
-            }
-
-            storage.write(Maps.transformValues(v1Records, recordPacker));
-            Iterator<EntityRecord> firstRevision = storage.readAll();
-            assertIteratorsEqual(v1Records.values()
-                                          .iterator(), firstRevision);
-
-            storage.write(Maps.transformValues(v2Records, recordPacker));
-            Iterator<EntityRecord> secondRevision = storage.readAll();
-            assertIteratorsEqual(v2Records.values()
-                                          .iterator(), secondRevision);
-        }
-    }
-
-    @Test
-    @DisplayName("fail to write visibility to non-existing record")
-    void notWriteStatusToNonExistent() {
-        I id = newId();
-        RecordStorage<I> storage = getStorage();
-
-        assertThrows(IllegalStateException.class,
-                     () -> storage.writeLifecycleFlags(id, archived()));
-    }
-
-    @Nested
-    @DisplayName("return visibility")
-    class ReturnVisibility {
-
-        @Test
-        @DisplayName("for missing record")
-        void forMissingRecord() {
-            I id = newId();
-            RecordStorage<I> storage = getStorage();
-            Optional<LifecycleFlags> optional = storage.readLifecycleFlags(id);
-            assertFalse(optional.isPresent());
-        }
-
-        @SuppressWarnings("ConstantConditions") // Converter nullability issues
-        @Test
-        @DisplayName("for new record")
-        void forNewRecord() {
-            I id = newId();
-            EntityRecord record = newStorageRecord(id);
-            RecordStorage<I> storage = getStorage();
-            storage.write(id, record);
-
-            Optional<LifecycleFlags> optional = storage.readLifecycleFlags(id);
-            assertTrue(optional.isPresent());
-            assertEquals(LifecycleFlags.getDefaultInstance(), optional.get());
-        }
-
-        @SuppressWarnings("OptionalGetWithoutIsPresent") // We verify in assertion.
-        @Test
-        @DisplayName("for record with updated visibility")
-        void forUpdatedRecord() {
-            I id = newId();
-            EntityRecord record = newStorageRecord(id);
-            RecordStorage<I> storage = getStorage();
-            storage.write(id, EntityRecordWithColumns.of(record));
-
-            storage.writeLifecycleFlags(id, archived());
-
-            Optional<LifecycleFlags> optional = storage.readLifecycleFlags(id);
-            assertTrue(optional.isPresent());
-            assertTrue(optional.get()
-                               .getArchived());
-        }
-    }
-
-    @Test
-    @DisplayName("accept records with empty storage fields")
-    void acceptRecordsWithEmptyColumns() {
-        I id = newId();
-        EntityRecord record = newStorageRecord(id);
-        EntityRecordWithColumns recordWithStorageFields = EntityRecordWithColumns.of(record);
-        assertFalse(recordWithStorageFields.hasColumns());
-        RecordStorage<I> storage = getStorage();
-
-        storage.write(id, recordWithStorageFields);
-        RecordReadRequest<I> readRequest = newReadRequest(id);
-        Optional<EntityRecord> actualRecord = storage.read(readRequest);
-        assertTrue(actualRecord.isPresent());
-        assertEquals(record, actualRecord.get());
+    private static void delete(TransactionalEntity<ProjectId, ?, ?> entity) {
+        TestTransaction.delete(entity);
     }
 
     @Test
     @DisplayName("write record with columns")
     void writeRecordWithColumns() {
-        I id = newId();
+        ProjectId id = newId();
         EntityRecord record = newStorageRecord(id);
-        TestCounterEntity<I> testEntity = new TestCounterEntity<>(id);
-        RecordStorage<I> storage = getStorage();
+        Entity<ProjectId, ?> testEntity = newEntity(id);
+        RecordStorage<ProjectId> storage = getStorage();
         EntityRecordWithColumns recordWithColumns = create(record, testEntity, storage);
         storage.write(id, recordWithColumns);
 
-        RecordReadRequest<I> readRequest = newReadRequest(id);
+        RecordReadRequest<ProjectId> readRequest = newReadRequest(id);
         Optional<EntityRecord> readRecord = storage.read(readRequest);
         assertTrue(readRecord.isPresent());
         assertEquals(record, readRecord.get());
@@ -441,7 +141,7 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
     @SuppressWarnings("OverlyLongMethod") // Complex test case (still tests a single operation)
     @Test
     @DisplayName("filter records by columns")
-    protected void filterByColumns() {
+    void filterByColumns() {
         Project.Status requiredValue = DONE;
         Int32Value wrappedValue = Int32Value
                 .newBuilder()
@@ -465,18 +165,19 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
                 .addFilter(aggregatingFilter)
                 .build();
 
-        RecordStorage<I> storage = getStorage();
+        RecordStorage<ProjectId> storage = getStorage();
 
-        EntityQuery<I> query = from(filters, storage);
-        I idMatching = newId();
-        I idWrong1 = newId();
-        I idWrong2 = newId();
+        EntityQuery<ProjectId> query = from(filters, storage);
+        ProjectId idMatching = newId();
+        ProjectId idWrong1 = newId();
+        ProjectId idWrong2 = newId();
 
-        TestCounterEntity<I> matchingEntity = new TestCounterEntity<>(idMatching);
-        TestCounterEntity<I> wrongEntity1 = new TestCounterEntity<>(idWrong1);
-        TestCounterEntity<I> wrongEntity2 = new TestCounterEntity<>(idWrong2);
+        TestCounterEntity matchingEntity = newEntity(idMatching);
+        TestCounterEntity wrongEntity1 = newEntity(idWrong1);
+        TestCounterEntity wrongEntity2 = newEntity(idWrong2);
 
         // 2 of 3 have required values
+
         matchingEntity.assignStatus(requiredValue);
         wrongEntity1.assignStatus(requiredValue);
         wrongEntity2.assignStatus(CANCELLED);
@@ -518,21 +219,21 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
 
     @Test
     @DisplayName("filter records by ID and not use columns")
-    protected void filterByIdAndNoColumns() {
+    void filterByIdAndNoColumns() {
         // Create the test data
-        I idMatching = newId();
-        I idWrong1 = newId();
-        I idWrong2 = newId();
+        ProjectId idMatching = newId();
+        ProjectId idWrong1 = newId();
+        ProjectId idWrong2 = newId();
 
-        TestCounterEntity<I> matchingEntity = new TestCounterEntity<>(idMatching);
-        TestCounterEntity<I> wrongEntity1 = new TestCounterEntity<>(idWrong1);
-        TestCounterEntity<I> wrongEntity2 = new TestCounterEntity<>(idWrong2);
+        Entity<ProjectId, ?> matchingEntity = newEntity(idMatching);
+        Entity<ProjectId, ?> wrongEntity1 = newEntity(idWrong1);
+        Entity<ProjectId, ?> wrongEntity2 = newEntity(idWrong2);
 
         EntityRecord fineRecord = newStorageRecord(idMatching, newState(idMatching));
         EntityRecord notFineRecord1 = newStorageRecord(idWrong1, newState(idWrong1));
         EntityRecord notFineRecord2 = newStorageRecord(idWrong2, newState(idWrong2));
 
-        RecordStorage<I> storage = getStorage();
+        RecordStorage<ProjectId> storage = getStorage();
 
         EntityRecordWithColumns recordRight = create(fineRecord, matchingEntity, storage);
         EntityRecordWithColumns recordWrong1 = create(notFineRecord1, wrongEntity1, storage);
@@ -557,7 +258,7 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
                 .newBuilder()
                 .setIdFilter(idFilter)
                 .build();
-        EntityQuery<I> query = from(filters, storage);
+        EntityQuery<ProjectId> query = from(filters, storage);
 
         // Perform the query
         Iterator<EntityRecord> readRecords =
@@ -569,18 +270,19 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
     @Test
     @DisplayName("read archived records if specified")
     protected void readArchivedRecords() {
-        I activeRecordId = newId();
-        I archivedRecordId = newId();
+        ProjectId activeRecordId = newId();
+        ProjectId archivedRecordId = newId();
 
         EntityRecord activeRecord =
                 newStorageRecord(activeRecordId, newState(activeRecordId));
         EntityRecord archivedRecord =
                 newStorageRecord(archivedRecordId, newState(archivedRecordId));
-        TestCounterEntity<I> activeEntity = new TestCounterEntity<>(activeRecordId);
-        TestCounterEntity<I> archivedEntity = new TestCounterEntity<>(archivedRecordId);
-        archivedEntity.archive();
+        TransactionalEntity<ProjectId, ?, ?> activeEntity = newEntity(activeRecordId);
+        TransactionalEntity<ProjectId, ?, ?>  archivedEntity = newEntity(archivedRecordId);
 
-        RecordStorage<I> storage = getStorage();
+        archive(archivedEntity);
+
+        RecordStorage<ProjectId> storage = getStorage();
         storage.write(activeRecordId, create(activeRecord, activeEntity, storage));
         storage.write(archivedRecordId, create(archivedRecord, archivedEntity, storage));
 
@@ -588,29 +290,30 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
                 .newBuilder()
                 .addFilter(all(eq(archived.toString(), true)))
                 .build();
-        EntityQuery<I> query = from(filters, storage);
+        EntityQuery<ProjectId> query = from(filters, storage);
         Iterator<EntityRecord> read = storage.readAll(query, FieldMask.getDefaultInstance());
         assertSingleRecord(archivedRecord, read);
     }
 
     @Test
     @DisplayName("filter archived or deleted records by ID on bulk read")
-    protected void filterByIdAndStatusInBulk() {
-        I activeId = newId();
-        I archivedId = newId();
-        I deletedId = newId();
+    void filterByIdAndStatusInBulk() {
+        ProjectId activeId = newId();
+        ProjectId archivedId = newId();
+        ProjectId deletedId = newId();
 
-        TestCounterEntity<I> activeEntity = new TestCounterEntity<>(activeId);
-        TestCounterEntity<I> archivedEntity = new TestCounterEntity<>(archivedId);
-        archivedEntity.archive();
-        TestCounterEntity<I> deletedEntity = new TestCounterEntity<>(deletedId);
-        deletedEntity.delete();
+        TransactionalEntity<ProjectId, ?, ?> activeEntity = newEntity(activeId);
+        TransactionalEntity<ProjectId, ?, ?> archivedEntity = newEntity(archivedId);
+        archive(activeEntity);
+
+        TransactionalEntity<ProjectId, ?, ?> deletedEntity = newEntity(deletedId);
+        delete(deletedEntity);
 
         EntityRecord activeRecord = newStorageRecord(activeId, activeEntity.getState());
         EntityRecord archivedRecord = newStorageRecord(archivedId, archivedEntity.getState());
         EntityRecord deletedRecord = newStorageRecord(deletedId, deletedEntity.getState());
 
-        RecordStorage<I> storage = getStorage();
+        RecordStorage<ProjectId> storage = getStorage();
         storage.write(deletedId, create(deletedRecord, deletedEntity, storage));
         storage.write(activeId, create(activeRecord, activeEntity, storage));
         storage.write(archivedId, create(archivedRecord, archivedEntity, storage));
@@ -620,23 +323,25 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
                 .addIds(toEntityId(archivedId))
                 .addIds(toEntityId(deletedId))
                 .build();
-        EntityFilters filters = EntityFilters.newBuilder()
-                                             .setIdFilter(idFilter)
-                                             .build();
-        EntityQuery<I> query = from(filters, storage).withLifecycleFlags(storage);
+        EntityFilters filters = EntityFilters
+                .newBuilder()
+                .setIdFilter(idFilter)
+                .build();
+        EntityQuery<ProjectId> query = from(filters, storage).withLifecycleFlags(storage);
         Iterator<EntityRecord> read = storage.readAll(query, FieldMask.getDefaultInstance());
         assertSingleRecord(activeRecord, read);
     }
 
     @Test
     @DisplayName("update entity column values")
-    protected void updateColumnValues() {
+    void updateColumnValues() {
         Project.Status initialStatus = DONE;
         @SuppressWarnings("UnnecessaryLocalVariable") // is used for documentation purposes.
                 Project.Status statusAfterUpdate = CANCELLED;
-        Int32Value initialStatusValue = Int32Value.newBuilder()
-                                                  .setValue(initialStatus.getNumber())
-                                                  .build();
+        Int32Value initialStatusValue = Int32Value
+                .newBuilder()
+                .setValue(initialStatus.getNumber())
+                .build();
         ColumnFilter status = eq("projectStatusValue", initialStatusValue);
         CompositeColumnFilter aggregatingFilter = CompositeColumnFilter
                 .newBuilder()
@@ -648,12 +353,13 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
                 .addFilter(aggregatingFilter)
                 .build();
 
-        RecordStorage<I> storage = getStorage();
+        RecordStorage<ProjectId> storage = getStorage();
 
-        EntityQuery<I> query = from(filters, storage);
+        EntityQuery<ProjectId> query = from(filters, storage);
 
-        I id = newId();
-        TestCounterEntity<I> entity = new TestCounterEntity<>(id);
+        ProjectId id = newId();
+        TestCounterEntity entity = newEntity(id);
+
         entity.assignStatus(initialStatus);
 
         EntityRecord record = newStorageRecord(id, newState(id));
@@ -668,6 +374,7 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
 
         // Update the entity columns of the record.
         entity.assignStatus(statusAfterUpdate);
+
         EntityRecordWithColumns updatedRecordWithColumns = create(record, entity, storage);
         storage.write(id, updatedRecordWithColumns);
 
@@ -677,17 +384,20 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
 
     @Test
     @DisplayName("filter records both by columns and IDs")
-    protected void filterByColumnsAndId() {
-        I targetId = newId();
-        TestCounterEntity<I> targetEntity = new TestCounterEntity<>(targetId);
-        TestCounterEntity<I> noMatchEntity = new TestCounterEntity<>(newId());
-        TestCounterEntity<I> noMatchIdEntity = new TestCounterEntity<>(newId());
-        TestCounterEntity<I> deletedEntity = new TestCounterEntity<>(newId());
+    void filterByColumnsAndId() {
+        ProjectId targetId = newId();
+        TestCounterEntity targetEntity = newEntity(targetId);
+        TestCounterEntity noMatchEntity = newEntity(newId());
+        TestCounterEntity noMatchIdEntity = newEntity(newId());
+        TestCounterEntity deletedEntity = newEntity(newId());
 
         targetEntity.assignStatus(CANCELLED);
         deletedEntity.assignStatus(CANCELLED);
-        deletedEntity.delete();
+
+        delete(deletedEntity);
+
         noMatchIdEntity.assignStatus(CANCELLED);
+
         noMatchEntity.assignStatus(DONE);
 
         write(targetEntity);
@@ -700,12 +410,13 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
                 .addIds(toEntityId(targetId))
                 .build();
         CompositeColumnFilter columnFilter = all(eq("projectStatusValue", CANCELLED.getNumber()));
-        EntityFilters filters = EntityFilters.newBuilder()
-                                             .setIdFilter(idFilter)
-                                             .addFilter(columnFilter)
-                                             .build();
-        RecordStorage<I> storage = getStorage();
-        EntityQuery<I> query = from(filters, storage).withLifecycleFlags(storage);
+        EntityFilters filters = EntityFilters
+                .newBuilder()
+                .setIdFilter(idFilter)
+                .addFilter(columnFilter)
+                .build();
+        RecordStorage<ProjectId> storage = getStorage();
+        EntityQuery<ProjectId> query = from(filters, storage).withLifecycleFlags(storage);
         Iterator<EntityRecord> read = storage.readAll(query, FieldMask.getDefaultInstance());
         List<EntityRecord> readRecords = newArrayList(read);
         assertEquals(1, readRecords.size());
@@ -720,7 +431,7 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
     @DisplayName("create unique states for same ID")
     void createUniqueStatesForSameId() {
         int checkCount = 10;
-        I id = newId();
+        ProjectId id = newId();
         Set<Message> states = newHashSet();
         for (int i = 0; i < checkCount; i++) {
             Message newState = newState(id);
@@ -738,22 +449,24 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
         Project.Status requiredValue = DONE;
         Project.Status value = Enum.valueOf(Project.Status.class, requiredValue.name());
         ColumnFilter status = eq(columnPath, value);
-        CompositeColumnFilter aggregatingFilter = CompositeColumnFilter.newBuilder()
-                                                                       .setOperator(ALL)
-                                                                       .addFilter(status)
-                                                                       .build();
-        EntityFilters filters = EntityFilters.newBuilder()
-                                             .addFilter(aggregatingFilter)
-                                             .build();
+        CompositeColumnFilter aggregatingFilter = CompositeColumnFilter
+                .newBuilder()
+                .setOperator(ALL)
+                .addFilter(status)
+                .build();
+        EntityFilters filters = EntityFilters
+                .newBuilder()
+                .addFilter(aggregatingFilter)
+                .build();
 
-        RecordStorage<I> storage = getStorage();
+        RecordStorage<ProjectId> storage = getStorage();
 
-        EntityQuery<I> query = from(filters, storage);
-        I idMatching = newId();
-        I idWrong = newId();
+        EntityQuery<ProjectId> query = from(filters, storage);
+        ProjectId idMatching = newId();
+        ProjectId idWrong = newId();
 
-        TestCounterEntity<I> matchingEntity = new TestCounterEntity<>(idMatching);
-        TestCounterEntity<I> wrongEntity = new TestCounterEntity<>(idWrong);
+        TestCounterEntity matchingEntity = newEntity(idMatching);
+        TestCounterEntity wrongEntity = newEntity(idWrong);
 
         matchingEntity.assignStatus(requiredValue);
         wrongEntity.assignStatus(CANCELLED);
@@ -790,12 +503,6 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
         return memoizedValue;
     }
 
-    private static EntityRecordWithColumns withRecordAndNoFields(EntityRecord record) {
-        return argThat(argument -> argument.getRecord()
-                                           .equals(record)
-                && !argument.hasColumns());
-    }
-
     private static void assertSingleRecord(EntityRecord expected, Iterator<EntityRecord> actual) {
         assertTrue(actual.hasNext());
         EntityRecord singleRecord = actual.next();
@@ -803,16 +510,17 @@ public abstract class RecordStorageTest<I, S extends RecordStorage<I>>
         assertEquals(expected, singleRecord);
     }
 
-    private EntityId toEntityId(I id) {
+    private static EntityId toEntityId(ProjectId id) {
         Any packed = Identifier.pack(id);
-        EntityId entityId = EntityId.newBuilder()
-                                    .setId(packed)
-                                    .build();
+        EntityId entityId = EntityId
+                .newBuilder()
+                .setId(packed)
+                .build();
         return entityId;
     }
 
-    private void write(Entity<I, ?> entity) {
-        RecordStorage<I> storage = getStorage();
+    private void write(Entity<ProjectId, ?> entity) {
+        RecordStorage<ProjectId> storage = getStorage();
         EntityRecord record = newStorageRecord(entity.getId(), entity.getState());
         storage.write(entity.getId(), create(record, entity, storage));
     }
