+++ server/src/test/java/io/spine/server/storage/memory/InMemoryRecordStorageTest.java
@@ -30,7 +30,6 @@ import io.spine.type.TypeUrl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static io.spine.base.Identifier.newUuid;
 import static io.spine.core.BoundedContextNames.newName;
 import static java.lang.String.format;
 import static java.lang.System.nanoTime;
@@ -41,33 +40,27 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
  */
 @DisplayName("InMemoryRecordStorage should")
 public class InMemoryRecordStorageTest
-        extends RecordStorageTest<ProjectId, InMemoryRecordStorage<ProjectId>> {
+        extends RecordStorageTest<InMemoryRecordStorage<ProjectId>> {
 
     @Override
     protected InMemoryRecordStorage<ProjectId> newStorage(Class<? extends Entity> cls) {
-        StorageSpec<ProjectId> spec = StorageSpec.of(newName(getClass().getSimpleName()),
-                                                     TypeUrl.of(Project.class),
-                                                     ProjectId.class);
+        StorageSpec<ProjectId> spec =
+                StorageSpec.of(newName(getClass().getSimpleName()),
+                               TypeUrl.of(Project.class),
+                               ProjectId.class);
         return InMemoryRecordStorage.newInstance(spec, false, cls);
-    }
-
-    @Override
-    protected ProjectId newId() {
-        ProjectId id = ProjectId.newBuilder()
-                                .setId(newUuid())
-                                .build();
-        return id;
     }
 
     @Override
     protected Message newState(ProjectId id) {
         String uniqueName = format("record-storage-test-%s-%s", id.getId(), nanoTime());
-        Project project = Project.newBuilder()
-                                 .setId(id)
-                                 .setStatus(Project.Status.CREATED)
-                                 .setName(uniqueName)
-                                 .addTask(Task.getDefaultInstance())
-                                 .build();
+        Project project = Project
+                .newBuilder()
+                .setId(id)
+                .setStatus(Project.Status.CREATED)
+                .setName(uniqueName)
+                .addTask(Task.getDefaultInstance())
+                .build();
         return project;
     }
